Code now compiles under solaris studio 12.4 with --std=c++11 option

+++ core/lib/GNSSEph/GloEphemeris.cpp
@@ -416,9 +416,9 @@ namespace gpstk
 
          // Get the Julian Day at 0 hours UT (jd)
       YDSTime ytime( time );
-      double year( ytime.year );
+      double year( static_cast<double>(ytime.year) );
       int doy( ytime.doy );
-      int temp( floor(365.25 * (year - 1.0)) + doy );
+      int temp( static_cast<int>(floor(365.25 * (year - 1.0))) + doy );
 
       double jd( static_cast<double>(temp)+ 1721409.5 );
